pyproject.toml
[project]
name = "simple-inference-server"
version = "0.1.0"
description = "A straightforward OpenAI-compatible inference API server for hosting multiple small models at the edge."
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [{name = "Your Team"}]
dependencies = [
  "fastapi>=0.123.4",
  "uvicorn[standard]>=0.38.0",
  "pydantic>=2.12.5",
  "numpy>=2.3.5",
  "torch>=2.9.1",
  "torchvision",
  "torchaudio",
  "transformers>=4.57.3",
  "accelerate>=1.12.0",
  "huggingface-hub>=0.36.0",
  "prometheus-client>=0.23.1",
  "pyyaml>=6.0.3",
  "python-dotenv>=1.2.1",
  "pillow>=12.0.0",
  "python-multipart>=0.0.20",
  "httpx>=0.28.1"
]

[project.optional-dependencies]
test = [
  "pytest>=9.0.1",
  "pytest-asyncio>=1.3.0",
  "httpx>=0.28.1",
  "anyio>=4.12.0"
]

[dependency-groups]
dev = [
  "pytest>=9.0.1",
  "pytest-asyncio>=1.3.0",
  "httpx>=0.28.1",
  "anyio>=4.12.0",
  "ruff>=0.14.7",
  "mypy>=1.19.0",
  "types-PyYAML>=6.0.12.20241230",
  "prometheus-client>=0.23.1",
]

[tool.uv.sources]
# Install PyTorch with CUDA support on Linux/Windows (CUDA doesn't exist for Mac).
# NOTE: We must explicitly request them as `dependencies` above. These improved
# versions will not be selected if they're only third-party dependencies.
torch = [
  { index = "pytorch-cuda", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchaudio = [
  { index = "pytorch-cuda", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cuda", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "pytorch-cuda"
# Use PyTorch built for NVIDIA Toolkit version 13.0.
# Available versions: https://pytorch.org/get-started/locally/
url = "https://download.pytorch.org/whl/cu130"
# Only use this index when explicitly requested by `tool.uv.sources`.
explicit = true

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 120
target-version = "py312"
src = ["app", "scripts", "tests"]
exclude = ["models"]  # HF cache dir

[tool.ruff.lint]
# Start from a sensible baseline and add a few high-signal plugins.
select = [
  "E4", "E7", "E9",  # Pycodestyle errors
  "F",               # Pyflakes
  "I",               # Import sorting (isort)
  "B",               # flake8-bugbear
  "UP",              # pyupgrade
  "SIM",             # flake8-simplify
  "PTH",             # flake8-use-pathlib
  "PL",              # pylint-inspired checks
  "S",               # flake8-bandit (security)
]
ignore = []

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]  # allow assert in tests

[tool.ruff.lint.isort]
combine-as-imports = true
known-first-party = ["app"]

[tool.mypy]
python_version = "3.12"
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_untyped_defs = true
check_untyped_defs = true
ignore_missing_imports = true
exclude = ["models"]

[[tool.mypy.overrides]]
module = [
  "torch.*",
  "transformers.*",
  "PIL.*",
  "numpy.*",
  "httpx.*",
  "starlette.*",
  "fastapi.*",
]
follow_imports = "skip"
ignore_missing_imports = true
